added principal to server

+++ final-capstone/java/src/main/java/com/techelevator/dao/UserDAO.java
@@ -12,7 +12,9 @@ public interface UserDAO {
 
     User findByUsername(String username);
 
-    int findIdByUsername(String username);
+    Long findIdByUsername(String username);
 
     boolean create(String username, String password, String role);
+
+
 }
