+++ final-capstone/java/src/main/java/com/techelevator/dao/UserSqlDAO.java
@@ -23,8 +23,8 @@ public class UserSqlDAO implements UserDAO {
     }
 
     @Override
-    public int findIdByUsername(String username) {
-        return jdbcTemplate.queryForObject("select user_id from users where username = ?", int.class, username);
+    public Long findIdByUsername(String username) {
+        return jdbcTemplate.queryForObject("select user_id from users where username = ?", Long.class, username);
     }
 
 	@Override
@@ -85,6 +85,7 @@ public class UserSqlDAO implements UserDAO {
 
         return userCreated;
     }
+
 
     private User mapRowToUser(SqlRowSet rs) {
         User user = new User();
